Moving validators to a separate file
